web/requirements.txt
flask
python-dotenv
psycopg2
requests
flask-cors
